requirements.txt
airtable-python-wrapper==0.15.3
python-dotenv==1.0.0 